更新配置文件

+++ src/main/java/com/syh/mall/vo/CommodityVO.java
@@ -36,7 +36,7 @@ public class CommodityVO {
 
         String name;
 
-        BigDecimal price;
+        Double price;
 
         Boolean status;
